+++ packages/components-providers/src/TrapStack/types.ts
@@ -26,7 +26,7 @@
 
 import { MutableRefObject } from 'react'
 
-export interface ManagerContextProps {
+export interface TrapStackContextProps {
   /**
    * Stores the element for the active scroll lock (null if none are active)
    */
@@ -55,18 +55,4 @@ export interface ManagerContextProps {
 
 export interface ElementMap {
   [key: string]: HTMLElement
-}
-
-export const getActiveElement = (elementMap: ElementMap) => {
-  // Sort the elements according to dom position and return the last
-  // which we assume to be stacked on top since all components using Portal
-  // share a single zIndexFloor and use dom order to determine stacking
-  const elements = Object.values(elementMap)
-  if (elements.length === 0) return null
-
-  const sortedElements = elements.sort((elementA, elementB) => {
-    const relationship = elementA.compareDocumentPosition(elementB)
-    return relationship > 3 ? 1 : -1
-  })
-  return sortedElements[0] || null
 }
